+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/project/ProjectDao.java
@@ -135,7 +135,7 @@ public class ProjectDao extends AbstractDao {
                 return null;
             }
 
-            Result<Record11<UUID, UUID, String, String, String, String, String, String, UUID, String, String>> repos = tx.select(
+            Result<Record12<UUID, UUID, String, String, String, String, String, Boolean, String, UUID, String, String>> repos = tx.select(
                     REPOSITORIES.REPO_ID,
                     REPOSITORIES.PROJECT_ID,
                     REPOSITORIES.REPO_NAME,
@@ -143,6 +143,7 @@ public class ProjectDao extends AbstractDao {
                     REPOSITORIES.REPO_BRANCH,
                     REPOSITORIES.REPO_COMMIT_ID,
                     REPOSITORIES.REPO_PATH,
+                    REPOSITORIES.IS_DISABLED,
                     REPOSITORIES.META.cast(String.class),
                     SECRETS.SECRET_ID,
                     SECRETS.SECRET_NAME,
@@ -153,7 +154,7 @@ public class ProjectDao extends AbstractDao {
                     .fetch();
 
             Map<String, RepositoryEntry> m = new HashMap<>();
-            for (Record11<UUID, UUID, String, String, String, String, String, String, UUID, String, String> repo : repos) {
+            for (Record12<UUID, UUID, String, String, String, String, String, Boolean, String, UUID, String, String> repo : repos) {
                 m.put(repo.get(REPOSITORIES.REPO_NAME),
                         new RepositoryEntry(
                                 repo.get(REPOSITORIES.REPO_ID),
@@ -163,6 +164,7 @@ public class ProjectDao extends AbstractDao {
                                 repo.get(REPOSITORIES.REPO_BRANCH),
                                 repo.get(REPOSITORIES.REPO_COMMIT_ID),
                                 repo.get(REPOSITORIES.REPO_PATH),
+                                repo.get(REPOSITORIES.IS_DISABLED),
                                 repo.get(SECRETS.SECRET_ID),
                                 repo.get(SECRETS.SECRET_NAME),
                                 repo.get(SECRETS.STORE_TYPE),